+++ biblioteca/src/main/java/com/residencia/biblioteca/services/EmprestimoService.java
@@ -29,8 +29,21 @@ public class EmprestimoService {
 		return emprestimoRepo.save(emprestimo);
 	}
 	
-	public void deletarEmprestimo(Emprestimo emprestimo) {
+	public Boolean deletarEmprestimo(Emprestimo emprestimo) {
+		if (emprestimo == null)
+			return false;
+
+		Emprestimo emprestimoExistente = buscarEmprestimoPorId(emprestimo.getCodigoEmprestimo());
+		if (emprestimoExistente == null)
+			return false;
+
 		emprestimoRepo.delete(emprestimo);
+
+		Emprestimo emprestimoContinuaExistindo = buscarEmprestimoPorId(emprestimo.getCodigoEmprestimo());
+		if (emprestimoContinuaExistindo == null)
+			return true;
+
+		return false;
 	}
 
 }
